requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-material-admin==1.8.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg==3.1.8
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
